% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_panel.R
\name{apply_panel}
\alias{apply_panel}
\title{Apply panel names to a flowframe or tibble}
\usage{
apply_panel(input, exp_info, panel = NULL)
}
\arguments{
\item{input}{A \code{\link[flowCore:flowFrame]{flowCore::flowFrame}} or tibble with the original channel names as the shortNames or columns (respectively).}

\item{exp_info}{A cytobank experiment info object generated by \code{\link{get_experimentinfo}}}

\item{panel}{A character string specifying the panel to be applied (neccessary if >1 panel present).}
}
\value{
If `input` was a flowFrame, returns a flowFrame with the 'desc' for each channel renamed approriately

If `input` was a tibble, return a tibble with the columns renamed approriately.
}
\description{
Apply panel names to a flowframe or tibble
}
