% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sampletags.R
\name{apply_sampletags}
\alias{apply_sampletags}
\title{Apply sampletags to the pData slot of a flowSet}
\usage{
apply_sampletags(input, exp_info, panel = NULL)
}
\arguments{
\item{input}{A \code{\link[flowCore:flowSet-class]{flowCore::flowSet}} or GatingSet}

\item{exp_info}{A cytobank experiment info object generated by \code{\link{get_experimentinfo}}}
}
\value{
a tagged flowSet or gatingSet, note that samples will be matched by filename to the FCS.filename slot of the samples tags
}
\description{
Apply sampletags to the pData slot of a flowSet
}
