% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scales.R
\name{apply_scales}
\alias{apply_scales}
\title{A simple wrapper for the population.list functions in cytobankAPI}
\usage{
apply_scales(input, exp_info, inverse = FALSE)
}
\arguments{
\item{exp_info}{cytobank experiment from fetchCytobankExperiment}
}
\value{
A list of populations and how they are defined by gates
}
\description{
A simple wrapper for the population.list functions in cytobankAPI
}
<<<<<<< HEAD
\seealso{
\code{\link[flowCore]{Subset}}

\code{\link[flowCore]{filter}}
}
=======
>>>>>>> 44fcb803f49ea4e40f952c6eb9c5aa71cc7c83e2
