% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe_flowFrame.R
\name{as.data.frame.flowFrame}
\alias{as.data.frame.flowFrame}
\title{S3 Method for the as.data.frame function to convert a flowFrame into a dataframe}
\usage{
\method{as.data.frame}{flowFrame}(x, ..., use_longnames = TRUE)
}
\arguments{
\item{x}{a flowFrame object}

\item{use_longnames}{if TRUE, uses the longName from the FCS file as the column names}
}
\value{
A data.frame of all the cells in the flowFrame
}
\description{
S3 Method for the as.data.frame function to convert a flowFrame into a dataframe
}
