% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe_flowSet.R
\name{as.data.frame.flowSet}
\alias{as.data.frame.flowSet}
\title{S3 Method for the as.data.frame function to convert a flowset into a dataframe}
\usage{
\method{as.data.frame}{flowSet}(x, ..., add_filename = TRUE,
  use_longnames = FALSE, add_pData = TRUE)
}
\arguments{
\item{x}{a flowSet object}

\item{add_filename}{if TRUE adds the filename of the original FCS file as a column}

\item{use_longnames}{if TRUE, uses the longName from the FCS file as the column names}
}
\value{
A data.frame of all the cells in the flowSet
}
\description{
S3 Method for the as.data.frame function to convert a flowset into a dataframe
}
