% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asinh_trans.R
\name{asinh_trans}
\alias{asinh_trans}
\title{A hyperbolic arcsinh transformation for use within ggplot2.}
\usage{
asinh_trans(cofactor)
}
\arguments{
\item{cofactor}{a cofactor to be used for the transformation}
}
\value{
A transformation of the axis by the cofactor, such that
y = asinh(x/cofactor)
}
\description{
A hyperbolic arcsinh transformation for use within ggplot2.
}
\examples{
library(ggplot2)

ggplot(diamonds, aes(x = price)) +
geom_histogram() +
scale_x_continuous(trans = asinh_trans(5))
}
\seealso{
\code{\link[scales]{pseudo_log_trans}} in the scales
package which implements a similar transformation
}
