% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_gates.R
\name{define_gates}
\alias{define_gates}
\title{A function that takes gates and scales returned by cytobank and generates flowcore compabitible gates}
\usage{
define_gates(gates, lut)
}
\arguments{
\item{gates}{returned by `get_gates`}

\item{lut, }{retruned by `get_expinfo`}
}
\value{
A flow frame containing the data which lie within the selected gate
}
\description{
A function that takes gates and scales returned by cytobank and generates flowcore compabitible gates
}
\seealso{
\code{\link[flowCore:Subset]{Subset}}

\code{\link[flowCore:filter]{filter}}
}
