% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCytobankExperiment.R
\name{fetchCytobankExperiment}
\alias{fetchCytobankExperiment}
\title{Imports Cytobank Experiment via the Cytobank API}
\usage{
fetchCytobankExperiment(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI:authentication]{authenticate}}}

\item{exp_id}{The experiment id of the experiment to be retrived}
}
\value{
A list of experiment information including

1. A tibble of FCS files contained in the experiment and their panels

2. A tibble of sample tags from the experiment

3. A table of scales exported directly from cytobank

4. A transformerList version of the scales for use in flowWorkspace

5. A list of compensation matricies

6. A gating + population heirarchy imported from Cytobank by CytoML

7. A list of panels contained in the experiment

8. The experiment ID (character vector)

9. The experiment name (character vector)
}
\description{
Imports Cytobank Experiment via the Cytobank API
}
