% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_population.R
\name{gate_population}
\alias{gate_population}
\title{A simple wrapper for the gating functions in flowCore}
\usage{
gate_population(flow_frame, population, poplist, gatelist, gate_defs, lut)
}
\arguments{
\item{flow_frame}{A flowframe containing the data to be gated}

\item{population}{Desired Population ("character")}

\item{poplist}{Population list}

\item{gatelist}{returned by `get_gates``}

\item{lut, }{retruned by `get_expinfo`}

\item{gate_def, }{returned by `define_gates`}
}
\value{
A flow frame containing the data which lie within the selected gate
}
\description{
A simple wrapper for the gating functions in flowCore
}
\seealso{
\code{\link[flowCore]{Subset}}

\code{\link[flowCore]{filter}}
}
