% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_in.R
\name{gatein}
\alias{gatein}
\title{A simple wrapper for the gating functions in flowCore}
\usage{
gatein(data, gate)
}
\arguments{
\item{data}{A flowframe containing the data to be gated}

\item{gate}{A flowCore gate object, as defined manually or by
\code{\link{define_gates}}}
}
\value{
A flow frame containing the data which lie within the selected gate
}
\description{
A simple wrapper for the gating functions in flowCore
}
\seealso{
\code{\link[flowCore:Subset]{Subset}}

\code{\link[flowCore:filter]{filter}}
}
