% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compensations.R
\name{get_compensations}
\alias{get_compensations}
\title{A simple function for dowloading compensation matricies from Cytobank}
\usage{
get_compensations(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI:authentication]{authenticate}}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A compensation matrix (if one exists).
}
\description{
A simple function for dowloading compensation matricies from Cytobank
}
\seealso{
\code{\link[CytobankAPI:compensations]{compensations.list}}
}
