% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcsfilelut.R
\name{get_fcsfilelut}
\alias{get_fcsfilelut}
\title{A simple wrapper for the population.list functions in cytobankAPI}
\usage{
get_fcsfilelut(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI:authentication]{autheticate}}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A list of populations and how they are defined by gates
}
\description{
A simple wrapper for the population.list functions in cytobankAPI
}
