% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gates.R
\name{get_gates}
\alias{get_gates}
\title{A simple function for dowloading gates from Cytobank.}
\usage{
get_gates(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI]}{autheticate}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A kist
}
\description{
A simple function for dowloading gates from Cytobank.
}
\seealso{
\code{\link[CytobankAPI]{compensations.list}}
}
