% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lut.R
\name{get_lut}
\alias{get_lut}
\title{A function for generating a scales and channel names look-up table from cytobank}
\usage{
get_lut(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI]}{autheticate}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A data.frame for looking up scale and channel parameters
}
\description{
A function for generating a scales and channel names look-up table from cytobank
}
