% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scales.R
\name{get_scales}
\alias{get_scales}
\title{A function for generating a scales and channel names look-up table from cytobank}
\usage{
get_scales(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI]}{autheticate}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A data.frame for looking up scale and channel parameters
}
\description{
A function for generating a scales and channel names look-up table from cytobank
}
