% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseArcsinhTransform.R
\name{inverseArcsinhTransform}
\alias{inverseArcsinhTransform}
\title{Creates the inverse of the asinh transformation, for backtransformations}
\usage{

  inverseArcsinhTransform(transformationId = "defaultInverseArcsinhTransform",
  a = 1, b = 1, c = 0)
}
\arguments{
\item{transformationId}{character string to identify the transformation}

\item{a}{positive double that corresponds to a shift about 0.}

\item{b}{positive double that corresponds to a scale factor.}

\item{c}{positive double}
}
\value{
Returns an object of class \code{transform}.
}
\description{
Creates the inverse of the asinh transformation, for backtransformations
}
\seealso{
\code{\link{transform-class}}, \code{\link{transform}},
\code{asinh}

Other Transform functions: \code{\link{inverseLogTransform}}
}
\concept{Transform functions}
\keyword{methods}
