% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertTransformerList.R
\name{invertTransformerList}
\alias{invertTransformerList}
\title{Inverts a TransformerList object to return the 'back transformation'}
\usage{
invertTransformerList(transforms)
}
\arguments{
\item{transforms}{a transformList object}
}
\value{
a transformList object
}
\description{
Inverts a TransformerList object to return the 'back transformation'
}
