% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_scales.R
\name{parse_scales}
\alias{parse_scales}
\title{Converts the scales table from cytobank to a transformlist for flowcore}
\usage{
parse_scales(scales_tbl, inverse = F)
}
\arguments{
\item{inverse, }{if true, returns the back-transformation to the original scales}

\item{exp_info}{object from get_experimentinfo}
}
\value{
a transformList object, with one slot for each channel
}
\description{
Converts the scales table from cytobank to a transformlist for flowcore
}
