% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_denisty_point_2d.R
\name{stat_density_point_2d}
\alias{stat_density_point_2d}
\title{Plots points while computing a kernel density estimate}
\usage{
stat_density_point_2d(mapping = NULL, data = NULL, geom = "point",
  position = "identity", ..., contour = FALSE, n = 100, h = NULL,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{n}{number of grid points in each direction}

\item{h}{Bandwidth (vector of length two). If `NULL`, estimated
using [MASS::bandwidth.nrd()].}
}
\description{
Plots points while computing a kernel density estimate
}
\section{Computed variables}{

\describe{
  \item{density}{the density estimate}
  \item{ndensity}{density estimate, scaled to maximum of 1}
}
}

