% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagFlowSet.R
\name{tagFlowSet}
\alias{tagFlowSet}
\title{Updates pData based on cytobank sampletags}
\usage{
tagFlowSet(myflowset, sampletags)
}
\arguments{
\item{myflowset}{a flowSet created by flowCore}

\item{sampletags}{a table of sampletags, containing a "FCS.Filename" column (at minimum)}
}
\value{
A flowSet with the pData slot updated based on the sample tags
}
\description{
Updates pData based on cytobank sampletags
}
