% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scales.R
\name{apply_scales}
\alias{apply_scales}
\title{A simple wrapper for the population.list functions in cytobankAPI}
\usage{
apply_scales(input, exp_info, inverse = FALSE)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI:authenticate]{CytobankAPI:authenticate}}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A list of populations and how they are defined by gates
}
\description{
A simple wrapper for the population.list functions in cytobankAPI
}
\seealso{
\code{\link[flowCore:Subset]{flowcore::Subset}}

\code{\link[flowCore:filter]{flowcore::filter}}
}
