% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCytobankExperiment.R
\name{fetchCytobankExperiment}
\alias{fetchCytobankExperiment}
\title{A wrapper for multiple functions that fetches experiment info from cytobank}
\usage{
fetchCytobankExperiment(cyto_session, exp_id)
}
\arguments{
\item{cyto_session}{A cytobank session created by \code{\link[CytobankAPI:authentication]{authenticate}}}

\item{exp_id}{The experiment ID, found in the URL of the experiment}
}
\value{
A list of experiment information including

1. A tibble of FCS files contained in the experiment from \code{\link{get_fcsfilelut}}

2. A tibble of sample tags from the experiment

3. A list of scales + panels with the corresponding including channel
 names, limits, and transformations generated by \code{\link{get_scales}}

4. A list of compensation matricies generated by \code{\link{get_compensations}}

5. A list of populations and their corresponding gates \code{\link{define_gates}}

6. A integer corresponding to the experiment ID
}
\description{
A wrapper for multiple functions that fetches experiment info from cytobank
}
