% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_population.R
\name{gate_population}
\alias{gate_population}
\title{A simple wrapper for the gating functions in flowCore}
\usage{
gate_population(flow_frame, population, exp_info, verbose = TRUE,
  apply_scales = TRUE)
}
\arguments{
\item{flow_frame}{A flowframe containing the data to be gated}

\item{population}{A character string corresponding to the desired population}

\item{exp_info}{A cytobank experiment info object generated by \code{\link{get_experimentinfo}}}

\item{verbose}{Should the gating hierarchy be printed during exectuion?
Useful for validaitng that the gating was applied correctly.}

\item{apply_scales}{Should the current scales be applied during gating?
Should be set to true unless the flowFrame has already been transformed.}
}
\value{
A flow frame containing the data which lie within the selected gates
}
\description{
A simple wrapper for the gating functions in flowCore
}
\seealso{
\code{\link[flowCore]{Subset}}

\code{\link[flowCore]{filter}}
}
